+++ tests/testparameterizedrepr.py
@@ -252,48 +252,3 @@ def test_pprint_signature_overriden():
     )
 
     assert t.param.pprint() == 'T()'
-
-
-# HTML repr
-
-class TestHTMLRepr:
-
-    @pytest.fixture
-    def Sub(self):
-        class Sub(param.Parameterized):
-            xsub = param.Parameter()
-
-        return Sub
-
-    @pytest.fixture
-    def P(self, Sub):
-        class P(param.Parameterized):
-            constant = param.Parameter(constant=True)
-            readonly = param.Parameter(readonly=True)
-            allow_None = param.Parameter(1, allow_None=False)
-            bounds = param.Number(bounds=(-10, 10))
-            objects_list = param.Selector(objects=[1, 2])
-            objects_dict = param.Selector(objects=dict(a=1, b=2))
-            sub = param.ClassSelector(default=Sub(xsub=1), class_=Sub)
-
-        return P
-
-    def test_repr_in_place(self):
-        assert hasattr(param.Parameterized.param, '_repr_html_')
-        assert hasattr(param.Parameterized().param, '_repr_html_')
-
-    def test_html_repr_class_str(self, P):
-        html = P.param._repr_html_()
-        assert isinstance(html, str)
-
-    def test_html_repr_inst_str(self, P):
-        html = P().param._repr_html_()
-        assert isinstance(html, str)
-
-    def test_html_repr_class_sub(self, P):
-        html = P.param._repr_html_()
-        assert '<details >\n <summary style="display:list-item; outline:none;">\n  <tt>Sub' in html
-
-    def test_html_repr_inst_sub(self, P):
-        html = P().param._repr_html_()
-        assert '<details >\n <summary style="display:list-item; outline:none;">\n  <tt>Sub' in html
